fix: fix log problem
